TERCERO/REDES2/PRACTICAS_V1/practica2/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
docopt==0.6.2
idna==2.9
pkg-resources==0.0.0
pycryptodome==3.9.7
requests==2.23.0
urllib3==1.25.9
